refactor CF context parsing functions
